Plans to support `Future` cancellation.
It is a common use case to schedule a large batch of operations against the driver. Not everything in that batch might be actually needed. In particular, there might be a error further down the line that makes the program to return early and `drop` the remainder of the scheduled batch. The driver should handle this gracefully by skipping the dropped operations.